requirements.txt
# GramSetu AI - Backend Dependencies (Production Ready)

# Core Flask framework
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
gunicorn==21.2.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Redis Cache
redis==5.0.1
hiredis==2.2.3

# AI/ML Libraries
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3

# OpenAI Integration
openai==1.6.1
langchain==0.1.0
langchain-openai==0.0.2

# Vector Database
pinecone-client==3.0.0

# Voice Processing & ASR
openai-whisper==20231117
pydub==0.25.1
SpeechRecognition==3.10.0
ffmpeg-python==0.2.0

# Language Detection
langdetect==1.0.9
indic-nlp-library==0.91

# Blockchain Integration
web3==6.13.0
thirdweb-sdk==3.5.0

# Security
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0

# Export Functionality
reportlab==4.0.7
openpyxl==3.1.2
Pillow==10.1.0

# Email
Flask-Mail==0.9.1

# Utilities
python-dateutil==2.8.2
requests==2.31.0
retrying==1.3.4

# Monitoring & Logging
python-json-logger==2.0.7

# Development & Testing
pytest==7.4.3
pytest-flask==1.2.0
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0


